requirements.txt
asgiref==3.6.0
Django==4.1.7
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
